refactor(02-native): split phase 2 screens into hooks and components
700-line profile-edit and 322-line ReminderCard were dumping logic,
API calls, and JSX into single files. Extracted into single-concern
units so screens are render-only going forward.

New files:
- src/lib/api.ts — shared API_BASE (was duplicated in 3 files)
- src/types/user.ts — shared UserProfile, Goal, GoalDraft types
- src/hooks/useProfile.ts — fetch + focus-refresh for profile screen
- src/hooks/useProfileEdit.ts — all edit state, handlers, API calls
- src/components/AvatarPickerModal.tsx — avatar bottom sheet modal
- src/components/ReminderPicker.tsx — time + day-of-week picker
