+++ app.json
@@ -1,0 +1,5 @@
+{
+  "android": {
+    "package": "com.anonymous.caliai"
+  }
+}
