+++ apps/native/app/_layout.tsx
@@ -10,6 +10,7 @@ const AppLayout = () => {
       <ThemeProvider>
         <SafeAreaProvider>
           <Stack>
+            <Stack.Screen name="(auth)" options={{ headerShown: false }} />
             <Stack.Screen name="(main)" options={{ headerShown: false }} />
           </Stack>
         </SafeAreaProvider>
