+++ apps/native/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "main": "index.js",
   "scripts": {
-    "dev": "expo start --web",
+    "dev": "expo start",
     "android": "expo run:android",
     "ios": "expo run:ios",
     "web": "expo start --web",
@@ -20,9 +20,9 @@
     "@repo/ui": "*",
     "expo": "^54.0.0",
     "expo-constants": "~18.0.8",
-    "expo-dev-client": "^55.0.20-canary-20260328-2049187",
-    "expo-image-manipulator": "^55.0.11",
-    "expo-image-picker": "^55.0.14",
+    "expo-dev-client": "~6.0.20",
+    "expo-image-manipulator": "~14.0.8",
+    "expo-image-picker": "~17.0.10",
     "expo-linear-gradient": "~15.0.7",
     "expo-linking": "~8.0.8",
     "expo-notifications": "~0.32.16",
@@ -32,7 +32,7 @@
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-hook-form": "^7.66.0",
-    "react-native": "0.81.4",
+    "react-native": "0.81.5",
     "react-native-fast-tflite": "^2.0.0",
     "react-native-nitro-modules": "^0.35.2",
     "react-native-safe-area-context": "~5.6.0",
