+++ apps/native/src/lib/api.ts
@@ -1,0 +1,1 @@
+export const API_BASE = "http://localhost:3000";
