+++ apps/native/tsconfig.json
@@ -2,10 +2,9 @@
   "extends": "expo/tsconfig.base",
   "compilerOptions": {
     "strict": true,
-    "baseUrl": "./",
     "paths": {
-      "@src/*": ["src/*"],
-      "@assets/*": ["assets/*"]
+      "@src/*": ["./src/*"],
+      "@assets/*": ["./assets/*"]
     },
     "types": ["react-native", "expo"]
   }
